moving common's states to shared module
