+++ src/app/portal/portal.module.ts
@@ -5,10 +5,6 @@ import { PortalRoutingModule } from './portal-routing.module';
 import { HomeComponent } from './components/home/home.component';
 import {SharedModule} from '../shared/shared.module';
 import { PortalComponent } from './portal.component';
-import {StoreModule} from '@ngrx/store';
-import {EffectsModule} from '@ngrx/effects';
-import {reducer} from './store/reducers/portal.reducers';
-import {PortalEffects} from './store/effects/portal.effects';
 
 
 @NgModule({
@@ -19,9 +15,7 @@ import {PortalEffects} from './store/effects/portal.effects';
   imports: [
     CommonModule,
     SharedModule,
-    PortalRoutingModule,
-    StoreModule.forFeature('portal', reducer),
-    EffectsModule.forFeature([PortalEffects])
+    PortalRoutingModule
   ]
 })
 export class PortalModule { }
