+++ src/app/portal/store/selectors/portal.selectors.ts
@@ -1,8 +1,8 @@
-import {State} from '../reducers/portal.reducers';
-import {createFeatureSelector, createSelector} from '@ngrx/store';
-
-export const fromPortal = createFeatureSelector('portal');
-
-export const selectLanguage = createSelector(fromPortal, (state: State) => state.language);
-
-
+// import {State} from '../reducers/portal.reducers';
+// import {createFeatureSelector, createSelector} from '@ngrx/store';
+//
+// export const fromPortal = createFeatureSelector('portal');
+//
+// export const selectLanguage = createSelector(fromPortal, (state: State) => state.language);
+//
+//
