+++ src/app/shared/shared.module.ts
@@ -4,6 +4,10 @@ import {HttpClient, HttpClientModule} from '@angular/common/http';
 import {TranslateHttpLoader} from '@ngx-translate/http-loader';
 import {TranslateLoader, TranslateModule} from '@ngx-translate/core';
 import {CommonService} from './services/common/common.service';
+import {StoreModule} from '@ngrx/store';
+import {reducer} from '../shared/store/reducers/common.reducers';
+import {EffectsModule} from '@ngrx/effects';
+import {CommonEffects} from './store/effects/common.effects';
 
 export function HttpLoaderFactory(http: HttpClient) {
   return new TranslateHttpLoader(http, './assets/i18n/', '.json');
@@ -20,7 +24,9 @@ export function HttpLoaderFactory(http: HttpClient) {
         useFactory: HttpLoaderFactory,
         deps: [HttpClient]
       }
-    })
+    }),
+    StoreModule.forFeature('common', reducer),
+    EffectsModule.forFeature([CommonEffects])
   ],
   exports: [
     CommonModule,
