+++ src/app/shared/store/selectors/common.selectors.ts
@@ -1,0 +1,8 @@
+import {CommonState} from '../reducers/common.reducers';
+import {createFeatureSelector, createSelector} from '@ngrx/store';
+
+export const fromPortal = createFeatureSelector('common');
+
+export const selectLanguage = createSelector(fromPortal, (state: CommonState) => state.language);
+
+
